% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/needs-package.R
\docType{package}
\name{needs-package}
\alias{needs-package}
\title{needs: easier package loading / auto-installation}
\description{
needs: easier package loading / auto-installation
}
\author{
Josh Katz
}
\references{
\url{http://www.github.com/joshkatz/needs}
}
\seealso{
\code{\link{needs}}
}

