% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/needs.R
\name{needs}
\alias{needs}
\title{Load/install packages.}
\usage{
needs(...)
}
\arguments{
\item{...}{Package names, as unquoted names or character strings.}
}
\description{
Attempts to attach packages. If it fails, first installs, then attaches. Use
it in place of \code{library} to automatically install, then attach, any
missing packages. Supply a minimum version number to update old packages as
needed.
}
\details{
\code{needs} can be used as all or part of an \code{\link{Rprofile}} or
sourced into R code directly, prior to any package loading.
}
\section{Versioning}{

Syntax for specifying minimum version numbers
}
\seealso{
\code{\link{needs-package}}
}

