% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prioritize.R
\name{prioritize}
\alias{prioritize}
\title{Re-attach packages to prevent masking.}
\usage{
prioritize(...)
}
\arguments{
\item{...}{Packages, given as unquoted names or character strings. Earlier arguments will be attached later (and therefore get a higher priority).}
}
\description{
\code{prioritize} detaches packages from the search path, then re-attaches them, placing them at the beginning of the search path to prevent masking.
}

