% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renderProfile.R
\name{renderProfile}
\alias{renderProfile}
\title{Generate Rprofile}
\usage{
renderProfile(append = T)
}
\arguments{
\item{append}{Whether to overwrite or append to the current Rprofile, if one exists. Defaults to true.}
}
\description{
Write to Rprofile
}
\details{
Writes to Rprofile in the current working directory.
}

