% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toProfile.R
\name{toProfile}
\alias{toProfile}
\title{Append to Rprofile in current working directory.}
\usage{
toProfile(dir = NULL, append = T)
}
\arguments{
\item{dir}{Target directory for Rprofile. Defaults to working directory.}

\item{append}{Whether to append to the current Rprofile, if one exists. Set
to false to overwrite.}
}
\description{
Exports the package functionality to .Rprofile in the current working
directory. Now the same code you run can be run on another system without
requiring any extra installation or throwing errors for uninstalled packages.
}
\examples{
\dontrun{
toProfile()
}

}

